create basic request and response
